ebox unlock is slow with many configs
Using e.g. pivy-box key unlock  on an ebox with many primary
configs can be very slow, as we may enumerate all PIV tokens
on the system for every single non-present device in the list
(which may take 1-2 sec sometimes).

With this change, we try agent keys first and then cache
enumeration results so that things take far less time.

On an ebox with 8 primary configs, where the config on the
system is the last one, this takes unlock from ~9sec to 0.25sec
with an agent or ~1sec without.