piv: ignore InvalidDataError when reading all certs.
The version libressl/openssh in use now is more strict about parsing
certificates than prior versions. This may cause "pivy-tool list" and other
tasks that need to iterate over card contents to be unable to list any
certificates on a card with one bad certificate. Changing read_all to _not_ abort
on InvalidDataError allows using other certs on the card, and also makes it
clear which cert is the problem.